requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
requests==2.33.1
python-dotenv==1.0.1
werkzeug==3.0.3
gunicorn==22.0.0
